requirements.txt
aiogram==3.1.1
aiosqlite==0.17.0
pandas==2.1.1
python-dotenv==1.0.0
matplotlib==3.8.0
tortoise-orm==0.20.0
uvloop==0.17.0
pandas==2.1.1